refactor: diffusion and sampler with a base module
