+++ src/opensynth/models/energydiff/_diffusion_base.py
@@ -1,0 +1,163 @@
+# Copyright Contributors to the Opensynth-energy Project.
+# SPDX-License-Identifier: Apache-2.0
+"""
+Name: diffusion base class and utility functions
+Author: Nan Lin (sentient-codebot)
+Date: Nov 2024
+
+"""
+import enum
+import math
+from abc import ABC, abstractmethod
+from collections import namedtuple
+from typing import Iterator
+
+import torch
+from torch import Tensor, nn
+from torch.nn.parameter import Parameter
+
+ModelPrediction = namedtuple(
+    "ModelPrediction", ["pred_noise", "pred_x_start", "pred_var_factor"]
+)
+
+
+class ModelMeanType(enum.Enum):
+    X_START = "x_start"
+    NOISE = "noise"
+    V = "v"
+
+
+class ModelVarianceType(enum.Enum):
+    FIXED_SMALL = "fixed_small"  # default, necessary for dpm-solver
+    FIXED_LARGE = "fixed_large"
+    LEARNED_RANGE = "learned_range"
+
+
+class LossType(enum.Enum):
+    MSE = "mse"
+    RESCALED_MSE = "rescaled_mse"
+    KL = "kl"
+    RESCALED_KL = "rescaled_kl"
+
+    def is_vb(self):
+        return self in (LossType.KL, LossType.RESCALED_KL)
+
+
+class BetaScheduleType(enum.Enum):
+    LINEAR = "linear"
+    COSINE = "cosine"  # default
+
+
+def extract(a: Tensor, t: Tensor, x_shape: torch.Size) -> Tensor:
+    """Extract values from tensor 'a' at positions specified by tensor 't' \
+        and reshape to 'x_shape'.
+
+    This function extracts values from tensor 'a' at positions specified \
+        by tensor 't',
+    and reshapes the result to match the shape specified by 'x_shape'.
+
+    Args:
+        a (Tensor): Source tensor to extract values from.
+        t (Tensor): Tensor containing positions where values should be \
+            extracted.
+        x_shape (torch.Size): Target shape for the output tensor.
+
+    Returns:
+        Tensor: Extracted values reshaped to match x_shape.
+
+    Note:
+        The function automatically converts tensor 'a' to float32 dtype and
+        matches the device of tensor 't'.
+    """
+    a = a.to(device=t.device, dtype=torch.float32)  # !dtype is fixed here
+    b, *_ = t.shape
+    dim_x = len(x_shape)
+    out = a.gather(-1, t)
+    target_shape = (b, *(1 for _ in range(dim_x - 1)))
+    out = out.reshape(*target_shape) + torch.zeros(x_shape, device=t.device)
+
+    return out
+
+
+def linear_beta_schedule(num_timestep: int) -> Tensor:
+    """get the linear beta schedule.
+
+    Arguments:
+        num_timestep -- int
+
+    Returns:
+        Tensor -- shape: (num_timestep,)
+    """
+    scale = 1000.0 / num_timestep
+    beta_start = scale * 1e-4
+    beta_end = scale * 2e-2
+    return torch.linspace(
+        beta_start, beta_end, num_timestep, dtype=torch.float64
+    )  # shape: (num_timestep,)
+
+
+def cosine_beta_schedule(num_timestep: int, s: float = 0.008):
+    "as proposed in https://openreview.net/forum?id=-NEXDKk8gZ"
+    num_step = num_timestep + 1
+    x = torch.linspace(
+        0, num_timestep, num_step, dtype=torch.float64
+    )  # shape: (num_step,)
+    alpha_cumprod = (
+        torch.cos((x / num_timestep + s) / (1 + s) * math.pi / 2) ** 2
+    )  # shape: (num_step,)
+    alpha_cumprod = alpha_cumprod / alpha_cumprod[0]  # shouldn't [0] be 1.?
+    beta_schedule = 1 - (
+        alpha_cumprod[1:] / alpha_cumprod[:-1]
+    )  # shape: (num_timestep,)
+    return torch.clip(beta_schedule, 0, 0.999)  # shape: (num_timestep,)
+
+
+def get_beta_schedule(beta_schedule_type: BetaScheduleType, num_timestep: int):
+    if beta_schedule_type == BetaScheduleType.LINEAR:
+        return linear_beta_schedule(num_timestep)
+    elif beta_schedule_type == BetaScheduleType.COSINE:
+        return cosine_beta_schedule(num_timestep)
+    else:
+        raise ValueError("type_beta_schedule must be one of linear, cosine")
+
+
+def get_loss_weight(
+    alpha_cumprod: Tensor, model_mean_type: ModelMeanType
+) -> Tensor:
+    # TODO unit test
+    # calculate loss weights
+    snr = alpha_cumprod / (1.0 - alpha_cumprod)
+    if model_mean_type == ModelMeanType.NOISE:
+        loss_weight = torch.ones_like(snr)
+    elif model_mean_type == ModelMeanType.X_START:
+        loss_weight = snr
+    elif model_mean_type == ModelMeanType.V:
+        loss_weight = snr / (snr + 1.0)
+
+    return loss_weight
+
+
+class DiffusionBase(ABC):
+    @property
+    @abstractmethod
+    def alpha_cumprod(self) -> Tensor:
+        pass
+
+    @property
+    @abstractmethod
+    def model_mean_type(self) -> ModelMeanType:
+        pass
+
+    @property
+    @abstractmethod
+    def model_variance_type(self) -> ModelVarianceType:
+        pass
+
+    @property
+    @abstractmethod
+    def model(self) -> nn.Module:
+        pass
+
+    @abstractmethod
+    def parameters(self) -> Iterator[Parameter]:
+        pass
