+++ src/opensynth/models/energydiff/diffusion.py
@@ -6,12 +6,8 @@ Author: Nan Lin (sentient-codebot)
 Date: Nov 2024
 
 """
-import enum
-import math
-from abc import ABC, abstractmethod
-from collections import namedtuple
 from functools import partial
-from typing import Callable, Iterator, Tuple
+from typing import Callable, Iterator
 
 import pytorch_lightning as pl
 import torch
@@ -19,162 +15,22 @@ import torch.nn.functional as F
 from einops import reduce
 from ema_pytorch import EMA
 from torch import Tensor, nn
-from torch.nn.parameter import Parameter
 from tqdm import tqdm
 
 from opensynth.data_modules.lcl_data_module import TrainingData
+from opensynth.models.energydiff._diffusion_base import (
+    BetaScheduleType,
+    DiffusionBase,
+    LossType,
+    ModelMeanType,
+    ModelPrediction,
+    ModelVarianceType,
+    extract,
+    get_beta_schedule,
+    get_loss_weight,
+)
 from opensynth.models.energydiff.model import DenoisingTransformer
-from opensynth.models.energydiff.sampler import (
-    DPM_Solver,
-    NoiseScheduleVP,
-    model_wrapper,
-)
-
-ModelPrediction = namedtuple(
-    "ModelPrediction", ["pred_noise", "pred_x_start", "pred_var_factor"]
-)
-
-
-class ModelMeanType(enum.Enum):
-    X_START = "x_start"
-    NOISE = "noise"
-    V = "v"
-
-
-class ModelVarianceType(enum.Enum):
-    FIXED_SMALL = "fixed_small"  # default, necessary for dpm-solver
-    FIXED_LARGE = "fixed_large"
-    LEARNED_RANGE = "learned_range"
-
-
-class LossType(enum.Enum):
-    MSE = "mse"
-    RESCALED_MSE = "rescaled_mse"
-    KL = "kl"
-    RESCALED_KL = "rescaled_kl"
-
-    def is_vb(self):
-        return self in (LossType.KL, LossType.RESCALED_KL)
-
-
-class BetaScheduleType(enum.Enum):
-    LINEAR = "linear"
-    COSINE = "cosine"  # default
-
-
-def extract(a: Tensor, t: Tensor, x_shape: torch.Size) -> Tensor:
-    """Extract values from tensor 'a' at positions specified by tensor 't' \
-        and reshape to 'x_shape'.
-
-    This function extracts values from tensor 'a' at positions specified \
-        by tensor 't',
-    and reshapes the result to match the shape specified by 'x_shape'.
-
-    Args:
-        a (Tensor): Source tensor to extract values from.
-        t (Tensor): Tensor containing positions where values should be \
-            extracted.
-        x_shape (torch.Size): Target shape for the output tensor.
-
-    Returns:
-        Tensor: Extracted values reshaped to match x_shape.
-
-    Note:
-        The function automatically converts tensor 'a' to float32 dtype and
-        matches the device of tensor 't'.
-    """
-    a = a.to(device=t.device, dtype=torch.float32)  # !dtype is fixed here
-    b, *_ = t.shape
-    dim_x = len(x_shape)
-    out = a.gather(-1, t)
-    target_shape = (b, *(1 for _ in range(dim_x - 1)))
-    out = out.reshape(*target_shape) + torch.zeros(x_shape, device=t.device)
-
-    return out
-
-
-def linear_beta_schedule(num_timestep: int) -> Tensor:
-    """get the linear beta schedule.
-
-    Arguments:
-        num_timestep -- int
-
-    Returns:
-        Tensor -- shape: (num_timestep,)
-    """
-    scale = 1000.0 / num_timestep
-    beta_start = scale * 1e-4
-    beta_end = scale * 2e-2
-    return torch.linspace(
-        beta_start, beta_end, num_timestep, dtype=torch.float64
-    )  # shape: (num_timestep,)
-
-
-def cosine_beta_schedule(num_timestep: int, s: float = 0.008):
-    "as proposed in https://openreview.net/forum?id=-NEXDKk8gZ"
-    num_step = num_timestep + 1
-    x = torch.linspace(
-        0, num_timestep, num_step, dtype=torch.float64
-    )  # shape: (num_step,)
-    alpha_cumprod = (
-        torch.cos((x / num_timestep + s) / (1 + s) * math.pi / 2) ** 2
-    )  # shape: (num_step,)
-    alpha_cumprod = alpha_cumprod / alpha_cumprod[0]  # shouldn't [0] be 1.?
-    beta_schedule = 1 - (
-        alpha_cumprod[1:] / alpha_cumprod[:-1]
-    )  # shape: (num_timestep,)
-    return torch.clip(beta_schedule, 0, 0.999)  # shape: (num_timestep,)
-
-
-def get_beta_schedule(beta_schedule_type: BetaScheduleType, num_timestep: int):
-    if beta_schedule_type == BetaScheduleType.LINEAR:
-        return linear_beta_schedule(num_timestep)
-    elif beta_schedule_type == BetaScheduleType.COSINE:
-        return cosine_beta_schedule(num_timestep)
-    else:
-        raise ValueError("type_beta_schedule must be one of linear, cosine")
-
-
-def get_loss_weight(
-    alpha_cumprod: Tensor, model_mean_type: ModelMeanType
-) -> Tensor:
-    # TODO unit test
-    # calculate loss weights
-    snr = alpha_cumprod / (1.0 - alpha_cumprod)
-    if model_mean_type == ModelMeanType.NOISE:
-        loss_weight = torch.ones_like(snr)
-    elif model_mean_type == ModelMeanType.X_START:
-        loss_weight = snr
-    elif model_mean_type == ModelMeanType.V:
-        loss_weight = snr / (snr + 1.0)
-
-    return loss_weight
-
-
-class DiffusionBase(ABC):
-    @property
-    @abstractmethod
-    def alpha_cumprod(self) -> Tensor:
-        pass
-
-    @property
-    @abstractmethod
-    def model_mean_type(self) -> ModelMeanType:
-        pass
-
-    @property
-    @abstractmethod
-    def model_variance_type(self) -> ModelVarianceType:
-        pass
-
-    @property
-    @abstractmethod
-    def model(self) -> nn.Module:
-        pass
-
-    @abstractmethod
-    def parameters(self) -> Iterator[Parameter]:
-        pass
+from opensynth.models.energydiff.sampler import DPMSolverSampler
 
 
 class GaussianDiffusion1D(nn.Module, DiffusionBase):
@@ -841,105 +697,6 @@ class GaussianDiffusion1D(nn.Module, DiffusionBase):
         if clip_denoised:
             all_sample = torch.clamp(all_sample, -1, 1)
         return all_sample
-
-
-class DPMSolverSampler:
-    def __init__(self, model: DiffusionBase, **kwargs):
-        super().__init__()
-        self.model = model
-        device = next(model.parameters()).device
-        to_torch = (
-            lambda x: x.clone().detach().to(torch.float32).to(device)
-        )  # noqa: E731
-        self.alphas_cumprod = to_torch(model.alpha_cumprod)
-
-    def register_buffer(self, name, attr):
-        if isinstance(attr, torch.Tensor):
-            if attr.device != next(self.model.parameters()).device:
-                attr = attr.to(next(self.model.parameters()).device)
-        setattr(self, name, attr)
-
-    @torch.no_grad()
-    def sample(
-        self,
-        S: int,  # steps
-        batch_size: int,  # N
-        shape: Tuple[int, int],  # (C, L)
-        x_T: Tensor | None = None,
-    ) -> Tuple[Tensor, None]:
-        """generate samples from the diffusion model using DPM-Solver.
-
-        Args:
-            S (int): number of steps to sample. has to be >= 1
-            batch_size (int): batch size
-            shape (tuple): shape of one sample, (num_channel, sequence_length)
-            x_T (Tensor, optional): initial tensor to be denoised. \
-                Defaults to None.
-
-        Returns:
-            Tuple[Tensor, None]: sampled tensor, None
-        """
-        # sampling
-        (C, L) = shape
-        size = (batch_size, C, L)
-
-        # print(f'Data shape for DPM-Solver sampling \
-        # is {size}, sampling steps {S}')
-
-        device = next(self.model.parameters()).device
-        if x_T is None:
-            img = torch.randn(size, device=device)
-        else:
-            img = x_T
-
-        ns = NoiseScheduleVP("discrete", alphas_cumprod=self.alphas_cumprod)
-
-        # original model function
-        def apply_model(x, t, c=None):
-            out = self.model.model.forward(x, t)
-            if (
-                self.model.model_variance_type
-                == ModelVarianceType.LEARNED_RANGE
-            ):  # GaussianDiffusion1D.model_variance_type
-                out = torch.split(out, out.shape[1] // 2, dim=1)[
-                    0
-                ]  # discard the learned variance
-            return out
-
-        # model mean type
-        if self.model.model_mean_type == ModelMeanType.NOISE:
-            model_type = "noise"
-        elif self.model.model_mean_type == ModelMeanType.X_START:
-            model_type = "x_start"
-        elif self.model.model_mean_type == ModelMeanType.V:
-            model_type = "v"
-        else:
-            raise ValueError(
-                f"Unknown model mean type {self.model.model_mean_type}"
-            )
-        model_fn = model_wrapper(
-            apply_model,
-            ns,
-            model_type=model_type,
-            guidance_type="classifier-free",
-            condition=None,
-            unconditional_condition=None,
-            cfg_scale=1.0,
-        )
-
-        dpm_solver = DPM_Solver(
-            model_fn, ns, predict_x0=True, thresholding=False
-        )
-        x = dpm_solver.sample(
-            img,
-            steps=S,
-            skip_type="time_uniform",
-            method="multistep",
-            order=3,
-            lower_order_final=True,
-        )
-
-        return x.to(device), None
 
 
 class PLDiffusion1D(pl.LightningModule):
